+++ imkit/imkit/amr/wavreader.c
@@ -151,14 +151,14 @@ int wav_get_header(void* obj, int* format, int* channels, int* sample_rate, int*
 
 int wav_read_data(void* obj, unsigned char* data, unsigned int length) {
 	struct wav_reader* wr = (struct wav_reader*) obj;
-	int n;
+	size_t n;
 	if (wr->wav == NULL)
 		return -1;
 	if (length > wr->data_length)
 		length = wr->data_length;
 	n = fread(data, 1, length, wr->wav);
 	wr->data_length -= length;
-	return n;
+	return (int)n;
 }
 
 int wav_duration(const char* fileName) {
