+++ imsdk/imsdk/AsyncTCP.m
@@ -89,7 +89,7 @@
         return;
     }
     const char *p = [self.data bytes];
-    int n = write_data(self.sock, (uint8_t*)p, self.data.length);
+    int n = write_data(self.sock, (uint8_t*)p, (int)self.data.length);
     if (n < 0) {
         NSLog(@"sock write error:%d", errno);
         dispatch_suspend(self.writeSource);
@@ -156,7 +156,7 @@
 #define BUF_SIZE (64*1024)
 -(void)onRead {
     while (1) {
-        int nread;
+        ssize_t nread;
         char buf[BUF_SIZE];
         
         do {
