+++ imsdk/imsdk/util.c
@@ -104,7 +104,7 @@ int sock_nonblock(int fd, int set) {
 }
 
 int write_data(int fd, uint8_t *bytes, int len) {
-    int n = 0;
+    ssize_t n = 0;
     
     do {
         n = send(fd, bytes, len, 0);
@@ -115,7 +115,7 @@ int write_data(int fd, uint8_t *bytes, int len) {
         }
         return 0;
     } else {
-        return n;
+        return (int)n;
     }
 }
